Update inversion circles
Separate inversion circles into two variations
Update some select warnings for julia and other vars
Make example for off center inversion circle

+++ packages/app/src/flame/variations/parametric/juliaScope.tsx
@@ -48,7 +48,7 @@ export const juliaScope = parametricVariation(
     const p3 = trunc(abs(p1) * random())
     const r = length(pos)
     const phi = atan2(pos.y, pos.x)
-    const lambda = select(-1.0, 1.0, random() > 0.5)
+    const lambda = f32(select(-1.0, 1.0, random() > 0.5))
     const t = (lambda * phi + 2 * PI.$ * p3) / p1
     const factor = pow(r, p2 / p1)
     return vec2f(cos(t), sin(t)).mul(factor)
